requirements.txt
requests
python-dotenv
plotly
pandas
pyperclip

flask
pywebio

twilio
adafruit-io
